websocket variant

+++ src/boards/boards.controller.ts
@@ -2,6 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import { BaseController } from "../common/base.controller";
 import { injectable, inject } from "inversify";
 import { join, extname } from 'path';
+import { existsSync, rm, unlink } from 'fs';
 import { TYPES } from "../types";
 import { ILogger } from "../logger/logger.interface";
 import { IConfigService } from "../config/config.service.interface";
@@ -10,7 +11,9 @@ import { AuthMiddleware } from "../common/auth.middleware";
 import { ValidatePropertiesMiddleware } from "../common/validateProperties.middleware";
 import fileUpload from "express-fileupload";
 import { v4 as uuidv4 } from 'uuid';
-import { unlink } from 'fs';
+import { ColumnsServices } from "../columns/columns.services";
+import { CardsServices } from "../cards/cards.services";
+import { UsersServices } from "../users/users.services";
 
 @injectable()
 export class BoardsController extends BaseController {
@@ -19,15 +22,21 @@ export class BoardsController extends BaseController {
         @inject(TYPES.ILogger) private loggerservice: ILogger,
         @inject(TYPES.ConfigService) private configService: IConfigService,
         @inject(TYPES.BoardsServices) private boardsService: BoardsServices,
+        @inject(TYPES.ColumnsServices) private columnsServices: ColumnsServices,
+        @inject(TYPES.UsersServices) private usersServices: UsersServices,
     ) {
         super(loggerservice);
         this.bindRoutes([
             { path: '/', method: 'get', func: this.getAllBoards, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
+            { path: '/guest', method: 'get', func: this.getAllGuestBoards, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
             { path: '/:boardId', method: 'get', func: this.getBoardInfo, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
             { path: '/', method: 'post', func: this.createBoard, middlewares: [new AuthMiddleware(this.configService.get('SECRET')), new ValidatePropertiesMiddleware(['name'])] },
-            { path: '/:boardId', method: 'put', func: this.updateBoardInfo, middlewares: [new AuthMiddleware(this.configService.get('SECRET')), new ValidatePropertiesMiddleware(['name'])] },
+            { path: '/:boardId', method: 'put', func: this.updateBoardInfo, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
             { path: '/:boardId', method: 'delete', func: this.deleteBoard, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
             { path: '/:boardId/columns', method: 'get', func: this.getColumnsByBoardId, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
+            { path: '/:boardId/participant/:userId', method: 'post', func: this.addParticipantToBoard, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
+            { path: '/:boardId/participants', method: 'get', func: this.getBoardParticipants, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
+            { path: '/:boardId/participant/:userId', method: 'delete', func: this.removeParticipantFromBoard, middlewares: [new AuthMiddleware(this.configService.get('SECRET'))] },
         ]);
     }
 
@@ -38,7 +47,7 @@ export class BoardsController extends BaseController {
                     const photo = files.photo as fileUpload.UploadedFile;
                     const photoId = uuidv4();
                     const uploadPath = join(__dirname, '../public/uploads', `boardPhoto_${photoId}${extname(photo.name)}`);
-    
+
                     photo.mv(uploadPath, async (err) => {
                         if (err) {
                             rej(null);
@@ -50,10 +59,10 @@ export class BoardsController extends BaseController {
                     rej(null);
                 }
             })
-            .then(async photoUrl => {
-                const board = await this.boardsService.createBoard(user.id, body.name, photoUrl);
-                this.ok(res, board);
-            })
+                .then(async photoUrl => {
+                    const board = await this.boardsService.createBoard(user.id, body.name, photoUrl);
+                    this.ok(res, board);
+                })
 
         } catch (error) {
             this.error(res, 500, 'Something was wrong, please try again later!');
@@ -63,6 +72,14 @@ export class BoardsController extends BaseController {
     async getAllBoards(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
             const boards = await this.boardsService.getAllBoards(req.user.id);
+            this.ok(res, boards);
+        } catch (error) {
+            this.error(res, 500, 'Something was wrong, please try again later!');
+        }
+    }
+    async getAllGuestBoards(req: Request, res: Response, next: NextFunction): Promise<void> {
+        try {
+            const boards = await this.boardsService.getAllGuestBoards(req.user.id);
             this.ok(res, boards);
         } catch (error) {
             this.error(res, 500, 'Something was wrong, please try again later!');
@@ -88,35 +105,45 @@ export class BoardsController extends BaseController {
         try {
             const boardId = parseInt(params.boardId, 10);
             const boardData = body;
+            if (!(boardData['name'] || files?.photo)) {
+                this.error(res, 422, 'The request must have at least one of the following properties: name, photo (file)');
+                return;
+            }
+
             const board = await this.boardsService.getBoardInfo(boardId);
 
-            if(!board) {
+            if (!board) {
                 this.error(res, 404, 'The board doesn\'t exist!');
                 return;
             }
 
             boardData.previewImage = await new Promise((rej) => {
                 if (files && Object.keys(files).length >= 1) {
-                    if(board.previewImage) unlink(join(__dirname, '../public', board.previewImage), (err) => err && this.loggerservice.error(err));
+                    if (board.previewImage) unlink(join(__dirname, '../public', board.previewImage), (err) => err && this.loggerservice.error(err));
 
                     const photo = files.photo as fileUpload.UploadedFile;
                     const photoId = uuidv4();
                     const uploadPath = join(__dirname, '../public/uploads', `boardPhoto_${photoId}${extname(photo.name)}`);
-    
+
                     photo.mv(uploadPath, async (err) => {
                         if (err) {
-                            rej(null);
+                            rej(undefined);
                         }
                         rej(`/uploads/boardPhoto_${photoId}${extname(photo.name)}`);
                     });
                 } else {
-                    rej(null);
+                    rej(undefined);
                 }
             })
-            const updatedBoard = await this.boardsService.updateBoardInfo(boardId, boardData);
+            const updatedBoard = await this.boardsService
+                .updateBoardInfo(
+                    boardId,
+                    { name: boardData.name, previewImage: boardData.previewImage }
+                );
 
             if (updatedBoard) {
                 this.ok(res, updatedBoard);
+                this.WsBoardController.uploadBoard(boardId, { type: 'ws/api/uploadBoard', body: updatedBoard })
             } else {
                 this.error(res, 404, 'Board is not found.');
             }
@@ -130,18 +157,36 @@ export class BoardsController extends BaseController {
             const boardId = parseInt(req.params.boardId, 10);
             const board = await this.boardsService.getBoardInfo(boardId);
 
-            if(!board) {
+            if (!board) {
                 this.error(res, 404, 'The board doesn\'t exist!');
                 return;
             }
-            const success = await this.boardsService.deleteBoard(boardId);
+            if (board.previewImage) unlink(join(__dirname, '../public', board.previewImage), (err) => err && this.loggerservice.error(err));
 
+            if(board.columns?.length) {
+                for(const column of board.columns) {
+                    const cards = await this.columnsServices.getCardsByColumnId(column.id);
+
+                    for(const card of cards) {
+                        const folderPath = join(__dirname, '../public/uploads/', `cardId_${card.cardId}`);
+                        if(existsSync(folderPath)) {
+                            rm(folderPath, { recursive: true, force: true }, (err) => {
+                                if(err) this.loggerservice.error(err);
+                            });
+                        }
+                    }
+                }
+            }
+
+            const success = await this.boardsService.deleteBoard(boardId);
             if (success) {
                 this.ok(res, { message: 'Board deleted successfully.' });
             } else {
                 this.error(res, 404, 'Board is not found.');
             }
         } catch (error) {
+            console.log(error);
+            
             this.error(res, 500, 'Something was wrong, please try again later!');
         }
     }
@@ -158,6 +203,82 @@ export class BoardsController extends BaseController {
 
             const columns = await this.boardsService.getColumnsByBoardId(boardId);
             this.ok(res, columns);
+        } catch (error) {
+            this.error(res, 500, 'Something went wrong, please try again later.');
+        }
+    }
+
+    async addParticipantToBoard(req: Request, res: Response, next: NextFunction): Promise<void> {
+        try {
+            const boardId = parseInt(req.params.boardId, 10);
+            const board = await this.boardsService.getBoardInfo(boardId);
+            if (!board) {
+                this.error(res, 404, 'Board is not found.');
+                return;
+            }
+
+            const userId = parseInt(req.params.userId, 10);
+            const user = await this.usersServices.getUserInfo(userId);
+            if(!user) {
+                this.error(res, 404, 'User is not found.');
+                return;
+            }
+            if(!board.participants) {
+                this.error(res, 500, 'Something went wrong, please try again later.');
+                return;
+            }
+            const result = await this.boardsService.addParticipant(board.participants.id, userId);
+            this.ok(res, result);
+        } catch (error) {
+            this.error(res, 500, 'Something went wrong, please try again later.');
+        }
+    }
+    async getBoardParticipants(req: Request, res: Response, next: NextFunction): Promise<void> {
+        try {
+            const boardId = parseInt(req.params.boardId, 10);
+            const board = await this.boardsService.getBoardInfo(boardId);
+
+            if (!board) {
+                this.error(res, 404, 'Board is not found.');
+                return;
+            }
+            const participants = await this.boardsService.getParticipants(boardId);
+            if(participants) {
+                this.ok(res, participants);
+            } else {
+                this.error(res, 404, 'Participants are not found');
+            }
+        } catch (error) {
+            this.error(res, 500, 'Something went wrong, please try again later.');
+        }
+    }
+    async removeParticipantFromBoard(req: Request, res: Response, next: NextFunction): Promise<void> {
+        try {
+            const boardId = parseInt(req.params.boardId, 10);
+            const board = await this.boardsService.getBoardInfo(boardId);
+            if (!board) {
+                this.error(res, 404, 'Board is not found.');
+                return;
+            }
+
+            const userId = parseInt(req.params.userId, 10);
+            const user = await this.usersServices.getUserInfo(userId);
+            if(!user) {
+                this.error(res, 404, 'User is not found.');
+                return;
+            }
+
+            if(board.createdBy.id === userId) {
+                this.error(res, 400, 'Can\'t remove board owner');
+                return;
+            }
+
+            if(!board.participants) {
+                this.error(res, 500, 'Something went wrong, please try again later.');
+                return;
+            }
+            const result = await this.boardsService.removeParticipant(board.participants.id, userId);
+            this.ok(res, result);
         } catch (error) {
             this.error(res, 500, 'Something went wrong, please try again later.');
         }
